E3c pre-move: lib→Analysis ref, credential relocation, FormatBaselineContext lift
Phase 0: PerformanceMonitor.Notifications now references PerformanceMonitor.Analysis
(one-way, no cycle) so the send services can move in later phases.
Phase 1: Dashboard SMTP/webhook credential statics relocated off EmailAlertService /
WebhookAlertService into a Dashboard-owned DashboardAlertCredentials helper; adapters,
SettingsWindow, and McpAlertTools repointed. Shared services will carry no CredentialService.
Phase 2: FormatBaselineContext + DayNames lifted from both Mcp/McpAnalysisTools.cs copies
into shared PerformanceMonitor.Analysis.BaselineContextFormatter; both AnalysisNotificationService
callers repointed.
