+++ Lite/Services/AnalysisNotificationService.cs
@@ -165,7 +165,7 @@ internal static class FindingMessageFormatter
 
         if (finding.RootFactMetadata is { Count: > 0 })
         {
-            var baseline = ToolRecommendations.FormatBaselineContext(finding.RootFactMetadata);
+            var baseline = BaselineContextFormatter.FormatBaselineContext(finding.RootFactMetadata);
             if (baseline is { Count: > 0 })
             {
                 var parts = baseline.Select(kv => $"{Humanize(kv.Key)} {kv.Value}");
